Add files via upload
the script searches for a specific email within the database, which could belong to either a “student” or an “instructor.” Once the email is found, the system generates a unique password reset token and sends it to the user’s email address. The link containing the token is valid for 60 seconds, ensuring a secure and time-limited password recovery process.